Get transcript sequences

This function is defunct. Use GenomicFeatures::extractTranscriptSeqs() instead.

Extract transcript sequences from a BSgenome object or an FaFile.

Usage

# S4 method for class 'GRangesList,BSgenome'
getTranscriptSeqs(query, subject, ...)
  # S4 method for class 'GRangesList,FaFile'
getTranscriptSeqs(query, subject, ...)
  # S4 method for class 'GRanges,FaFile'
getTranscriptSeqs(query, subject, ...)

Arguments

query

A GRangesList object containing exons or cds grouped by transcript.

subject

A BSgenome object or a FaFile from which the sequences will be taken.

...

Additional arguments

Details

getTranscriptSeqs is a wrapper for the extractTranscriptSeqs and getSeq functions. The purpose is to allow sequence extraction from either a BSgenome or FaFile. Transcript sequences are extracted based on the boundaries of the feature provided in the query (i.e., either exons or cds regions).

Value

A DNAStringSet instance containing the sequences for all transcripts specified in query.

Author

Valerie Obenchain

See also

Examples

## See ?extractTranscriptSeqs in the GenomicFeatures package.