indexVcf {VariantAnnotation}	R Documentation

Create index files for VCF files

Description

indexVcf() creates an index file for compressed VCF files, if the index file does not exist.

Usage

## S4 method for signature 'character'
indexVcf(x, ...)
## S4 method for signature 'VcfFile'
indexVcf(x, ...)
## S4 method for signature 'VcfFileList'
indexVcf(x, ...)

Arguments

x
character(), VcfFile, or VcfFileList pointing to bgzf-compressed VCF files.

...
Additional arguments to indexTabix

Details

If x is a character vector, assumes they are the path(s) to bgzf-compressed VCF file(s). If an index does not exist, one is created. VCF files can be compreseed using bgzip. A VcfFile or VcfFileList is returned.

If a VcfFile or VcfFileList is given, the index file is checked, if it does not exist it will crete one. If the index file was NA or missing, the path of the associated VCF file is used as the index file path. An updated VcfFile or VcfFileList is returned.

Value

VcfFile or VcfFileList

Author(s)

Lori Shepherd

See Also

VcfFile

Examples

fl <- system.file(
    "extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
    mustWork=TRUE
)
vcf1 <- indexVcf(fl)
vcf1

[Package VariantAnnotation version 1.36.0 Index]