seqinfo {VariantAnnotation}	R Documentation

Get seqinfo for VCF file

Description

Get seqinfo for VCF file

Usage

## S4 method for signature 'VcfFile'
seqinfo(x)
## S4 method for signature 'VcfFileList'
seqinfo(x)

Arguments

x
Either character(), VcfFile, or VcfFileList

Details

If a VcfFileThe file header is scanned an appropriate seqinfo object in given. If a VcfFileList is given, all file headers are scanned, and appropriate combined seqinfo object is given.

Value

Seqinfo object

Author(s)

Lori Shepherd

See Also

VcfFile, Seqinfo

Examples

fl <- system.file("extdata", "chr7-sub.vcf.gz", package="VariantAnnotation",
                  mustWork=TRUE)
vcf <- VcfFile(fl)
seqinfo(vcf)

[Package VariantAnnotation version 1.36.0 Index]