VRangesList-class {VariantAnnotation}	R Documentation

VRangesList objects

Description

VRangesList is a virtual class representing a list of VRanges objects and should behave much like any other derivative of List. It has both a simple and compressed implementation. VRangesList provides conveniences for manipulating sets of VRanges objects.

Constructor

VRangesList(...): Creates a VRangesList object from VRanges objects in ....

Accessors

alt(x): Returns a CharacterList or RleList, effectively by calling alt(x[[i]]) on each element of x.

ref(x): Returns a CharacterList, effectively by calling ref(x[[i]]) on each element of x.

Utilities

stackSamples(x): Concentrates the elements in x, using names(x) to appropriately fill sampleNames in the result.

Author(s)

Michael Lawrence

Examples

## construction
example(VRanges)
vrl <- VRangesList(sampleA = vr, sampleB = vr)
stackSamples(vrl)

[Package VariantAnnotation version 1.36.0 Index]